ed25519/requirements.txt
qrcode[pil]~=8.2
pyzbar~=0.1.9
Pillow~=11.3
cryptography~=43.0